+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
--
+- Added battery monitoring of BLE devices by @Flo100.
 ### Changed
 - Filesystem no longer updates when auto-update is unchecked. 
 - Holding shifter buttons on boot now erases LittleFS as well as resetting settings.